+++ src/models/Team.ts
@@ -1,0 +1,9 @@
+export default interface Team {
+  abbreviation: string,
+  city: string,
+  conference: string,
+  division: string,
+  full_name: string,
+  id: number,
+  name: string,
+}